+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/manager/environment/NymEnvironmentManager.kt
@@ -21,10 +21,6 @@ class NymEnvironmentManager @Inject constructor(
 		return getFeatureFlags()?.isDomainFrontingEnabled() ?: false
 	}
 
-	override suspend fun isQuicEnabled(): Boolean {
-		return getFeatureFlags()?.isQuicEnabled() ?: false
-	}
-
 	private suspend fun getFeatureFlags(): FeatureFlags? {
 		return try {
 			backendManager.getFeatureFlags()