+++ nym-vpn-app/src/screens/InitialNavigation.tsx
@@ -6,7 +6,7 @@ import { routes } from '../router';
 let navigationHandled = false;
 
 function InitialNavigation() {
-  const { account, initialized } = useMainState();
+  const { account, initialized, daemonStatus } = useMainState();
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -20,7 +20,7 @@ function InitialNavigation() {
     if (!account) {
       navigate(routes.onboarding, { replace: true });
     }
-  }, [account, initialized, navigate]);
+  }, [account, daemonStatus, initialized, navigate]);
 
   return null;
 }